+++ term_image/image/common.py
@@ -17,7 +17,7 @@ from functools import wraps
 from operator import gt, mul, sub
 from random import randint
 from types import FunctionType, TracebackType
-from typing import Any, Optional, Tuple, Union
+from typing import Any, Optional, Set, Tuple, Union
 from urllib.parse import urlparse
 
 import PIL
@@ -26,7 +26,7 @@ from PIL import Image, UnidentifiedImageError
 
 from .. import get_font_ratio
 from ..exceptions import InvalidSize, TermImageException, URLNotFoundError
-from ..utils import get_terminal_size, no_redecorate
+from ..utils import ClassInstanceMethod, get_terminal_size, no_redecorate
 
 _ALPHA_THRESHOLD = 40 / 255  # Default alpha threshold
 _FORMAT_SPEC = re.compile(
@@ -131,6 +131,7 @@ class BaseImage(ABC):
     # Data Attributes
 
     _supported = None
+    _render_methods: Set[str] = {}
 
     # Special Methods
 
@@ -731,7 +732,8 @@ class BaseImage(ABC):
     @abstractmethod
     def is_supported(cls) -> bool:
         """Returns ``True`` if the render style or graphics protocol implemented by
-        the invoking class is supported by the terminal. Otherwise, ``False``.
+        the invoking class is supported by the :term:`active terminal`.
+        Otherwise, ``False``.
         """
         raise NotImplementedError
 
@@ -756,6 +758,66 @@ class BaseImage(ABC):
             )
         if self._is_animated:
             self._seek_position = pos
+
+    @ClassInstanceMethod
+    def set_render_method(self_or_cls, method: Optional[str] = None) -> bool:
+        """Sets the render method used by the instances of subclasses providing
+        multiple render methods.
+
+        Args:
+            method: The render method to be set or ``None`` for a reset.
+
+        Returns:
+
+            * ``False``, if the given method is not recognized/implmented by the
+              calling class (or class of the calling instance).
+            * ``True``, if the given method is implemented and has been set or
+              *method* is ``None``.
+
+        Raises:
+            TypeError: *method* is not a string or ``None``.
+
+        See the **Render Methods** section in the description of the subclasses that
+        implement such for their specific usage.
+
+        If called via:
+
+           - the class, sets the class-wide render method.
+           - an instance, sets the instance-specifc render method.
+
+        If *method* is ``None`` and this method is called via:
+
+           - the class, the class-wide render method is reset to the default.
+           - an instance, the instance-specific render method is removed, so that it
+             uses the class-wide render method thenceforth.
+
+        Any instance without a specific render method set, uses the class-wide render
+        method.
+
+        Any class not implementing multiple render methods simply returns ``False``
+        for any value of *method* other than ``None``, so it's safe to blindly try to
+        set a render method for any subclass on another subclass.
+        """
+        if method is not None and not isinstance(method, str):
+            raise TypeError(
+                f"'method' must be a string or `None` (got: {type(method).__name__!r})"
+            )
+
+        if None is not method not in self_or_cls._render_methods:
+            return False
+
+        if not method:
+            if isinstance(self_or_cls, __class__):
+                try:
+                    del self_or_cls._render_method
+                except AttributeError:
+                    pass
+            elif self_or_cls._render_methods:
+                self_or_cls._render_method = self_or_cls._default_render_method
+        else:
+            self_or_cls._render_method = method
+
+        return True
 
     def set_size(
         self,
@@ -1020,6 +1082,14 @@ class BaseImage(ABC):
             raise ValueError(f"Scale value out of range (got: {value})")
         return value
 
+    @classmethod
+    def _clear_images(self, *args, **kwargs) -> bool:
+        """Used by some graphics-protocol-based styles to clear images on-screen.
+
+        Any overriding method should return ``True``.
+        """
+        return False
+
     def _display_animated(
         self,
         img: PIL.Image.Image,
@@ -1046,6 +1116,7 @@ class BaseImage(ABC):
             duration = self._frame_duration
             start = time.time()
             for frame in image_it._animate(img, alpha, fmt):
+                self._clear_images()
                 print(frame, end="", flush=True)  # Current frame
 
                 # Left-over of current frame's duration
@@ -1131,11 +1202,16 @@ class BaseImage(ABC):
         alpha: Union[None, float, str],
         *,
         size: Optional[Tuple[int, int]] = None,
-    ) -> Tuple[PIL.Image.Image, Tuple[Tuple[int, int, int]], Tuple[int]]:
+        pixel_data: bool = True,
+    ) -> Tuple[
+        PIL.Image.Image, Optional[Tuple[Tuple[int, int, int]]], Optional[Tuple[int]]
+    ]:
         """Returns the PIL image instance and pixel data required to render an image.
 
-        If *size* is given (in pixels), it is used instead of the pixel-equivalent of
-        the image size (or auto size, if size is unset).
+        Args:
+            size: If given (in pixels), it is used instead of the pixel-equivalent of
+              the image size (or auto size, if size is unset).
+            pixel_data: If ``False``, ``None`` is returned for all pixel data.
 
         The returned image is appropriately converted, resized and composited
         (if need be).
@@ -1157,32 +1233,31 @@ class BaseImage(ABC):
                 img = img.convert("RGB").resize((width, height))
             except ValueError:
                 raise ValueError("Image size or scale too small") from None
-            rgb = tuple(img.getdata())
-            a = (255,) * (width * height)
-            alpha = None
+            if pixel_data:
+                rgb = tuple(img.getdata())
+                a = (255,) * (width * height)
         else:
             try:
                 img = img.convert("RGBA").resize((width, height))
             except ValueError:
                 raise ValueError("Image size or scale too small") from None
+
             if isinstance(alpha, str):
                 bg = Image.new("RGBA", img.size, alpha)
                 bg.alpha_composite(img)
                 if img is not self._source:
                     img.close()
-                img = bg
-                alpha = None
-            rgb = tuple(img.convert("RGB").getdata())
-            if alpha is None:
-                a = (255,) * (width * height)
-            else:
+                img = bg.convert("RGB")
+                if pixel_data:
+                    a = (255,) * (width * height)
+            elif pixel_data:
                 alpha = round(alpha * 255)
                 a = [0 if val < alpha else val for val in img.getdata(3)]
-                # To distinguish `0.0` from `None` in truth value tests
-                if alpha == 0.0:
-                    alpha = True
 
-        return img, rgb, a
+            if pixel_data:
+                rgb = tuple(img.convert("RGB").getdata())
+
+        return (img, *(pixel_data and (rgb, a) or (None, None)))
 
     @abstractmethod
     def _get_render_size(self) -> Tuple[int, int]: